+++ packages/plugin-packages/mashroom-security-provider-ldap/test/MashroomLdapSecurityProvider.test.ts
@@ -26,6 +26,26 @@ describe('MashroomLdapSecurityProvider', () => {
         expect(redirectUrl).toBe('/login?redirectUrl=%2Ffoo%2Fbar');
     });
 
+    it('cancels the authentication process if the original url is the login page itself', async () => {
+        let redirectUrl = null;
+
+        const req: any = {
+            originalUrl: '/login?redirectUrl=https://foo.com'
+        };
+        const res: any = {
+            redirect: (url: string) => redirectUrl = url
+        };
+
+        const ldapClient: any = {};
+        const simpleSecurityProvider = new MashroomLdapSecurityProvider('/login', '', '', null, null,
+            null, null, ldapClient, '', 1800, loggerFactory);
+
+        const result = await simpleSecurityProvider.authenticate(req, res);
+
+        expect(result).toBeTruthy();
+        expect(result.status).toBe('error');
+    });
+
     it('passes the authentication hints to the login page', async () => {
         let redirectUrl = null;